mlb-ai-chatbot-extension/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.3.0
langchain-core==0.3.29
langchain-google-genai==2.0.1
langsmith==0.1.125
langchain-community==0.3.0
google-generativeai==0.8.0
python-dotenv==1.0.1
requests==2.31.0
orjson==3.10.1
loguru==0.7.2
pydantic==2.7.4